ps.c

+++ trap.c
@@ -8,6 +8,9 @@
 #include "traps.h"
 #include "spinlock.h"
 
+//Added a declaration of mappages to this file, disabled in vm.cstatic int
+int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);
+
 // Interrupt descriptor table (shared by all CPUs).
 struct gatedesc idt[256];
 extern uint vectors[];  // in vectors.S: array of 256 entry pointers
@@ -75,7 +78,7 @@ trap(struct trapframe *tf)
   case T_IRQ0 + IRQ_SPURIOUS:
     cprintf("cpu%d: spurious interrupt at %x:%x\n",
             cpuid(), tf->cs, tf->eip);
-    lapiceoi();
+ 		lapiceoi();
     break;
 
   //PAGEBREAK: 13
@@ -86,7 +89,21 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+	unsigned int fault_address = rcr2();	
+	if(tf->trapno == T_PGFLT)
+	{
+		char *mem;
+		uint a = PGROUNDDOWN(fault_address);	
+		mem = kalloc();
+		memset(mem, 0, PGSIZE);
+		//cprintf("[Virtual Address: %x][myproc->sz: %d][mem: %x]\n",a, myproc()->sz,mem);
+		mappages(myproc()->pgdir, (char*)a, PGSIZE, V2P(mem), PTE_W|PTE_U);
+  		break;
+	}
+
+//  End of my addition
     // In user space, assume process misbehaved.
+
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
             myproc()->pid, myproc()->name, tf->trapno,
